Remove unnecessary `IntPtr` cast now `nint` is integral type
Follow-up to https://github.com/dotnet/coreclr/pull/15947/commits/7aa2dba47466fd006350ac0261b192a041ba5a15

+++ src/libraries/System.Private.CoreLib/src/System/Buffer.cs
@@ -149,10 +149,8 @@ namespace System
                 goto BuffersOverlap;
             }
 
-            // Use "(IntPtr)(nint)len" to avoid overflow checking on the explicit cast to IntPtr
-
-            ref byte srcEnd = ref Unsafe.Add(ref src, (IntPtr)(nint)len);
-            ref byte destEnd = ref Unsafe.Add(ref dest, (IntPtr)(nint)len);
+            ref byte srcEnd = ref Unsafe.Add(ref src, (nint)len);
+            ref byte destEnd = ref Unsafe.Add(ref dest, (nint)len);
 
             if (len <= 16)
                 goto MCPY02;
